requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.17.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0